[hotfix][coordination] Move utility methods of SlotManager to SlotManagerUtils for deduplication
