+++ flink-runtime/src/main/java/org/apache/flink/runtime/resourcemanager/slotmanager/SlotManagerUtils.java
@@ -1,0 +1,36 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.runtime.resourcemanager.slotmanager;
+
+import org.apache.flink.runtime.clusterframework.types.ResourceProfile;
+import org.apache.flink.runtime.resourcemanager.WorkerResourceSpec;
+
+/** Utilities for {@link SlotManager} implementations. */
+public class SlotManagerUtils {
+    public static ResourceProfile generateDefaultSlotResourceProfile(
+            WorkerResourceSpec workerResourceSpec, int numSlotsPerWorker) {
+        return ResourceProfile.newBuilder()
+                .setCpuCores(workerResourceSpec.getCpuCores().divide(numSlotsPerWorker))
+                .setTaskHeapMemory(workerResourceSpec.getTaskHeapSize().divide(numSlotsPerWorker))
+                .setTaskOffHeapMemory(
+                        workerResourceSpec.getTaskOffHeapSize().divide(numSlotsPerWorker))
+                .setManagedMemory(workerResourceSpec.getManagedMemSize().divide(numSlotsPerWorker))
+                .setNetworkMemory(workerResourceSpec.getNetworkMemSize().divide(numSlotsPerWorker))
+                .build();
+    }
+}
